fix(server): store timestamp in milliseconds (#222)
* fix(server): store timestamp in milliseconds


* fix: convert type to long


* fix: update swagger schema


* fix: remove conversion to int


---------

+++ GomokuServer/src/GomokuServer.Application/Games/Dto/ClockDto.cs
@@ -1,11 +1,11 @@
 ﻿namespace GomokuServer.Application.Games.Dto;
 
 [TranspilationSource]
-public record ClockDto(int Black, int White)
+public record ClockDto(long Black, long White)
 {
 	[Required]
-	public int Black { get; } = Black;
+	public long Black { get; } = Black;
 
 	[Required]
-	public int White { get; } = White;
+	public long White { get; } = White;
 }
